@@ -1,37 +1,44 @@
 {
-    "name": "qotp-demo",
+    "name": "demo",
     "version": "0.0.1",
     "description": "A Quasar Project",
-    "productName": "QOtp Demo",
+    "productName": "q-otp-input demo",
     "author": "RUBisco0211 <1205456072@qq.com>",
+    "type": "module",
     "private": true,
     "scripts": {
         "test": "echo \"No test specified\" && exit 0",
         "dev": "quasar dev",
-        "build": "quasar build"
+        "build": "quasar build",
+        "postinstall": "quasar prepare",
+        "update-qotp": "quasar ext add @rubisco0211/qotp@latest"
     },
     "dependencies": {
-        "@highlightjs/vue-plugin": "^2.1.0",
         "@quasar/extras": "^1.16.4",
-        "highlight.js": "^10.7.3",
-        "mavon-editor": "^3.0.1",
-        "pinia": "^2.3.1",
+        "@vueuse/core": "^12.7.0",
+        "markdown-it": "^14.1.0",
+        "markdown-it-anchor": "^9.2.0",
+        "markdown-it-prism": "^2.3.0",
+        "pinia": "^3.0.1",
         "pinia-plugin-persistedstate": "^4.2.0",
-        "quasar": "^2.6.0",
-        "vue": "^3.0.0",
-        "vue-hljs": "^3.0.1",
-        "vue-i18n": "^10.0.5",
-        "vue-router": "^4.0.0"
+        "quasar": "^2.16.0",
+        "unplugin-vue-markdown": "^28.3.1",
+        "vite-plugin-markdown-preview": "^1.1.1",
+        "vue": "^3.4.18",
+        "vue-i18n": "^9.2.2",
+        "vue-router": "^4.0.12"
     },
     "devDependencies": {
-        "@quasar/app-vite": "^1.3.0",
-        "@rubisco0211/quasar-app-extension-qotp": "^0.0.6",
-        "@types/node": "^12.20.21",
+        "@intlify/unplugin-vue-i18n": "^2.0.0",
+        "@quasar/app-vite": "^2.1.0",
+        "@rubisco0211/quasar-app-extension-qotp": "^0.0.21",
+        "@types/node": "^20.17.19",
         "autoprefixer": "^10.4.2",
-        "typescript": "^4.5.4"
+        "shiki": "^3.0.0",
+        "typescript": "~5.5.3"
     },
     "engines": {
-        "node": "^18 || ^16 || ^14.19",
+        "node": "^28 || ^26 || ^24 || ^22 || ^20 || ^18",
         "npm": ">= 6.13.4",
         "yarn": ">= 1.21.1"
     }
