@@ -1,22 +1,41 @@
-import { boot } from "quasar/wrappers";
+import { defineBoot } from "#q-app/wrappers";
 import { Directive } from "vue";
-
-import hljs from "highlight.js";
+import { codeToHtml } from "shiki";
+import { copyToClipboard, Dark } from "quasar";
 
 const directives: Record<string, Directive> = {
-    hljs: (el: HTMLElement, binding: { value: { lang: string } }) => {
+    highlight: async (
+        el: HTMLElement,
+        binding: { value: { lang: string; code: string } }
+    ) => {
         const {
-            value: { lang },
+            value: { lang, code },
         } = binding;
-        const blocks = el.querySelectorAll("pre code");
-        blocks.forEach((block) => {
-            block.classList.add(lang);
-            hljs.highlightBlock(block as HTMLElement);
+        el.classList.add("q-pa-md");
+        const html = await codeToHtml(code, {
+            lang,
+            theme: Dark.isActive ? "github-dark" : "github-light",
         });
+        el.innerHTML = html;
+    },
+    clickCopy: (
+        el: HTMLElement,
+        binding: {
+            value: {
+                content: string;
+                onSuccess?: () => void;
+                onError?: () => void;
+            };
+        }
+    ) => {
+        const { content, onSuccess, onError } = binding.value;
+        el.onclick = async () => {
+            copyToClipboard(content).then(onSuccess).catch(onError);
+        };
     },
 };
 
-export default boot(({ app }) => {
+export default defineBoot(({ app }) => {
     Object.entries(directives).forEach(([name, directive]) => {
         app.directive(name, directive);
     });
