@@ -1,17 +1,33 @@
-import { boot } from "quasar/wrappers";
+import { defineBoot } from "#q-app/wrappers";
 import { createI18n } from "vue-i18n";
-import enUS from "src/i18n/en-US";
-import zhCN from "src/i18n/zh-CN";
 
-const i18n = createI18n({
-    locale: "en",
-    fallbackLocale: "zh",
-    messages: {
-        en: enUS,
-        zh: zhCN,
-    },
-});
+import messages from "src/i18n/index";
 
-export default boot(({ app }) => {
+export type MessageLanguages = keyof typeof messages;
+// Type-define 'en-US' as the master schema for the resource
+export type MessageSchema = (typeof messages)["en-US"];
+
+// See https://vue-i18n.intlify.dev/guide/advanced/typescript.html#global-resource-schema-type-definition
+/* eslint-disable @typescript-eslint/no-empty-object-type */
+declare module "vue-i18n" {
+    // define the locale messages schema
+    export interface DefineLocaleMessage extends MessageSchema {}
+
+    // define the datetime format schema
+    export interface DefineDateTimeFormat {}
+
+    // define the number format schema
+    export interface DefineNumberFormat {}
+}
+/* eslint-enable @typescript-eslint/no-empty-object-type */
+
+export default defineBoot(({ app }) => {
+    const i18n = createI18n<{ message: MessageSchema }, MessageLanguages>({
+        locale: "en-US",
+        legacy: false,
+        messages,
+    });
+
+    // Set i18n instance on app
     app.use(i18n);
 });
