@@ -1,18 +1,20 @@
-import { RouteRecordRaw } from 'vue-router';
+import type { RouteRecordRaw } from "vue-router";
 
 const routes: RouteRecordRaw[] = [
-  {
-    path: '/',
-    component: () => import('layouts/MainLayout.vue'),
-    children: [{ path: '', component: () => import('pages/IndexPage.vue') }],
-  },
+    {
+        path: "/",
+        component: () => import("src/layouts/MainLayout.vue"),
+        children: [
+            { path: "", component: () => import("src/pages/IndexPage.vue") },
+        ],
+    },
 
-  // Always leave this as last one,
-  // but you can also remove it
-  {
-    path: '/:catchAll(.*)*',
-    component: () => import('pages/ErrorNotFound.vue'),
-  },
+    // Always leave this as last one,
+    // but you can also remove it
+    {
+        path: "/:catchAll(.*)*",
+        component: () => import("src/pages/ErrorNotFound.vue"),
+    },
 ];
 
 export default routes;
