@@ -1,16 +1,15 @@
-import { store } from "quasar/wrappers";
+import { defineStore } from "#q-app/wrappers";
 import { createPinia } from "pinia";
-import { Router } from "vue-router";
 import { createPersistedState } from "pinia-plugin-persistedstate";
-
 /*
  * When adding new properties to stores, you should also
  * extend the `PiniaCustomProperties` interface.
  * @see https://pinia.vuejs.org/core-concepts/plugins.html#typing-new-store-properties
  */
 declare module "pinia" {
+    // eslint-disable-next-line @typescript-eslint/no-empty-object-type
     export interface PiniaCustomProperties {
-        readonly router: Router;
+        // add your custom properties here, if any
     }
 }
 
@@ -23,9 +22,11 @@ declare module "pinia" {
  * with the Store instance.
  */
 
-export default store((/* { ssrContext } */) => {
+export default defineStore((/* { ssrContext } */) => {
     const pinia = createPinia();
-    // 默认自动持久化到 localStorage；配置全局key格式
+
+    // You can add Pinia plugins here
+    // pinia.use(SomePiniaPlugin)
     pinia.use(
         createPersistedState({
             storage: localStorage,
